checking the existence of files (fix #18)
The existence of all documents from any source (tlpdb or texdocs) are checked
just before adding the document to the candidate list: if a file does not exist
it will not add to the candidate list, so it will never tried to be viewed nor
be shown in the result list. Even no scoring calculation will be executed to
save people's time and resource.

New debug category "search" will report this behavior.