use normname instead of shortname for scoring (fix #38 & fix #49)
Texdoc used to use so-called shortnames, each of which is a filename and only a
parent directory. This makes it difficult to adjust for some queries, e.g., #38
and #49. Now it uses normnames for the scoring. Accordingly, some adjustments
are added to the texdoc.cnf. In addition, give small minus adjustments for the
latex-dev stuff so that the stable version of kernel-related documents are
prioritized.