Send users online when they have no local docs.
Many users choose to install TeX Live without any documentation. This
means that `texdoc` is completely unhelpful, since it only searches
locally. This commit makes it so that `texdoc` detects if *any* local
documentation exists; if it does, it searches as normal; if not, it
informs the user and offers to open online documentation for them.